+++ Scripts/Extensions/AudioExtensions.cs
@@ -7,14 +7,21 @@ namespace Etienne
         public static AudioSource SetSoundToSource(this AudioSource source, Sound sound)
         {
             source.outputAudioMixerGroup = sound.Parameters.AudioMixerGroup;
+            source.priority = sound.Parameters.Priority;
             source.clip = sound.Clip;
-            source.loop = sound.Parameters.Loop;
+            source.panStereo = sound.Parameters.StereoPan;
             source.pitch = sound.Parameters.Pitch;
             source.volume = sound.Parameters.Volume;
             source.spatialBlend = sound.Parameters.SpacialBlend;
             source.minDistance = sound.Parameters.MinDistance;
             source.maxDistance = sound.Parameters.MaxDistance;
             return source;
+        }
+
+        public static void StopLoop(this AudioSource source)
+        {
+            source.Stop();
+            Pools.AudioSourcePool.Instance.Enqueue(source);
         }
     }
 }